Citation update: agol19 -> agol20
Previously, "exoplanet:agol19" pointed to the arXiv reference for
"Analytic Planetary Transit Light Curves and Derivatives for Stars with
Polynomial Limb Darkening'.

Since this paper is now in press at AJ, the reference should be updated.
This will help people who copy-paste the citation into their papers, so
that they don't need to fix the reference during the proofing stages ;-)